backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-filter==25.2
uvicorn==0.40.0
psycopg2-binary==2.9.11